refactor: consolidate do and dowhile tests into a single test file
